+++ app/src/main/java/com/example/fragmentnotes/ui/dialogs/DialogExitApp.java
@@ -1,4 +1,4 @@
-package com.example.fragmentnotes.ui;
+package com.example.fragmentnotes.ui.dialogs;
 
 import android.app.Dialog;
 import android.content.Context;
