+++ app/src/main/java/com/example/fragmentnotes/ui/recycler/NotesAdapter.java
@@ -1,4 +1,4 @@
-package com.example.fragmentnotes.ui;
+package com.example.fragmentnotes.ui.recycler;
 
 import android.view.View;
 import android.view.ViewGroup;
