+++ app/src/main/res/layout/fragment_about.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.AboutFragment"
+    tools:context=".ui.additioanlFragments.AboutFragment"
     android:background="@color/background_color_about_fragment">
 
     <TextView
