+++ app/src/main/res/layout/fragment_profile.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ProfileFragment"
+    tools:context=".ui.additioanlFragments.ProfileFragment"
     android:background="@color/background_color_profile_fragment">
 
     <TextView
