+++ app/src/main/res/layout/fragment_settings.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SettingsFragment"
+    tools:context=".ui.additioanlFragments.SettingsFragment"
     android:background="@color/background_color_settings_fragment">
 
     <TextView
